+++ R/AnnotationMethods.R
@@ -36,7 +36,7 @@
 #' exampleTSSr <- annotateCluster(
 #'     exampleTSSr, clusters = "consensusClusters", filterCluster = TRUE
 #' )
-#' head(slot(exampleTSSr, "assignedClusters")[["control"]])
+#' head(assignedClusters(exampleTSSr, sample = "control"))
 #'
 setGeneric("annotateCluster", function(
   object, clusters = "consensusClusters",