+++ R/ExpressionMethods.R
@@ -26,7 +26,11 @@
 #'     exampleTSSr, comparePairs = list(c("control", "treat")),
 #'     pval = 0.01, useMultiCore = FALSE
 #' )
-#' head(slot(exampleTSSr, "DEtables")[["control_VS_treat"]])
+#' head(DEtables(
+#'     exampleTSSr,
+#'     comparison = "control_VS_treat",
+#'     result = "all"
+#' ))
 setGeneric(
     "deGene",
     function(object, comparePairs = list(c("control", "treat")), pval = 0.01,