+++ R/ImportMethods.R
@@ -40,7 +40,7 @@
 #'     mergeIndex = 1
 #' )
 #' importedTSSr <- getTSS(importedTSSr)
-#' head(slot(importedTSSr, "TSSrawMatrix"))
+#' head(TSSmatrix(importedTSSr, data = "raw"))
 setGeneric("getTSS", function(
   object,
   sequencingQualityThreshold = 10,