+++ R/ShiftingMethods.R
@@ -24,7 +24,7 @@
 #' exampleTSSr <- shiftPromoter(
 #'     exampleTSSr, comparePairs = list(c("control", "treat")), pval = 0.01
 #' )
-#' head(slot(exampleTSSr, "PromoterShift")[["control_VS_treat"]])
+#' head(PromoterShift(exampleTSSr, comparison = "control_VS_treat"))
 setGeneric(
     "shiftPromoter",
     function(object, comparePairs, pval = 0.01) standardGeneric("shiftPromoter"),